Refactor database functions from run function
